requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow==10.0.1
celery==5.3.4
redis==5.0.1